# lib/api-zod/src/generated/types/memorySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deriv AI Market Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface MemorySummary {
  totalPatterns: number;
  avgAccuracy: number;
  topPatterns: string[];
  recentLessons: string[];
}