# lib/api-zod/src/generated/types/predictionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deriv AI Market Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionInputDirection } from './predictionInputDirection';
import type { PredictionInputIndicators } from './predictionInputIndicators';

export interface PredictionInput {
  symbol: string;
  direction: PredictionInputDirection;
  confidence: number;
  entryPrice: number;
  /** @nullable */
  analysisId?: number | null;
  indicators: PredictionInputIndicators;
}